feat(hooks): supplier information hook
